requirements.txt
# Combined requirements - NTMods (sentiment/indicator/trust) + NTWeb
# Only packages actually used in the codebase (imports verified)

beautifulsoup4>=4.13.4
feedparser>=6.0.11
Flask>=3.1.2
numpy>=2.2.5
pandas>=2.3.3
python-dateutil>=2.9.0
python-dotenv>=1.2.1
requests>=2.32.3
ta
vaderSentiment>=3.3.2
yfinance>=1.1.0
